Fix code organization TODO
