+++ packages/kit/src/runtime/internal/utils.js
@@ -2,3 +2,14 @@ export function find_anchor(node) {
 	while (node && node.nodeName.toUpperCase() !== 'A') node = node.parentNode; // SVG <a> elements have a lowercase name
 	return node;
 }
+
+export function get_base_uri(window_document) {
+	let baseURI = window_document.baseURI;
+
+	if (!baseURI) {
+		const baseTags = window_document.getElementsByTagName('base');
+		baseURI = baseTags.length ? baseTags[0].href : window_document.URL;
+	}
+
+	return baseURI;
+}
